updated mesh files for eddie
