+++ eddie_base_description/package.xml
@@ -1,20 +1,31 @@
 <?xml version="1.0"?>
 <package format="2">
   <name>eddie_base_description</name>
-  <version>0.1.0</version>
-  <description>Description of the base of Eddie in form of URDF files and meshes</description>
+  <version>0.0.0</version>
+  <description>The eddie_base_description package</description>
 
-  <maintainer email="robotics@inf.h-brs.de">MAS robotics</maintainer>
+  <!-- One maintainer tag required, multiple allowed, one person per tag -->
+  <!-- Example:  -->
+  <!-- <maintainer email="jane.doe@example.com">Jane Doe</maintainer> -->
+  <maintainer email="sawantk@todo.todo">sawantk</maintainer>
 
+  <!-- One license tag required, multiple allowed, one license per tag -->
+  <!-- Commonly used license strings: -->
+  <!--   BSD, MIT, Boost Software License, GPLv2, GPLv3, LGPLv2.1, LGPLv3 -->
   <license>GPLv3</license>
 
-  <author email="aleksandar.mitrevski@h-brs.de">Alex Mitrevski</author>
-  <author email="dharmingeo@gmail.com">Dharmin B.</author>
-  <author email="sushant.chavan@smail.inf.h-brs.de">Sushant Chavan</author>
-
+  <!-- Dependencies -->
   <buildtool_depend>ament_cmake</buildtool_depend>
+  <!-- <depend>xacro</depend> -->
+  <!-- <depend>robot_state_publisher</depend> -->
+  <!-- <depend>rviz2</depend> -->
+  <!-- <depend>geometry_msgs</depend> -->
+  <!-- <depend>urdf</depend> -->
+  <exec_depend>ros2launch</exec_depend>
 
+  <!-- Export build information -->
   <export>
     <build_type>ament_cmake</build_type>
   </export>
+
 </package>
