--- eddie_base_description/urdf/eddie_base_gz.urdf.xacro
@@ -1,97 +1,0 @@
-<?xml version="1.0"?>
-<robot xmlns:xacro="http://www.ros.org/wiki/xacro">
-
-    <!-- simplified box collision geometry for base -->
-    <xacro:property name="robot_name" value="eddie"/>
-
-    <xacro:property name="base_size_x" value="0.635"/>
-    <xacro:property name="base_size_y" value="0.582"/>
-    <xacro:property name="base_size_z_bottom" value="0.12"/>
-    <xacro:property name="wheel_slit_height" value="0.1"/>
-    <xacro:property name="laser_slit_height" value="0.05"/>
-    <xacro:property name="base_size_z_top" value="0.1"/>
-    <xacro:property name="base_size_z" value="${base_size_z_bottom + laser_slit_height + base_size_z_top}"/>
-    <xacro:property name="base_mass" value="40.0"/>
-
-    <xacro:include filename="$(find eddie_base_description)/urdf/kelo_wheel_gz.urdf.xacro" />
-
-    <xacro:macro name="eddie_base" params="name parent *origin use_gpu_2d_laser:=true">
-
-        <!-- BASE -->
-        <link name="${name}_link">
-            <visual>
-                <origin xyz="0 0 ${wheel_slit_height + base_size_z_bottom/2}" rpy="0 0 0"/>
-                <geometry>
-                    <box size="${base_size_x} ${base_size_y} ${base_size_z_bottom}" />
-                </geometry>
-                <material name="DarkGrey"/>
-            </visual>
-            <visual>
-                <origin xyz="0 0 ${wheel_slit_height + base_size_z_bottom + laser_slit_height/2}" rpy="0 0 0"/>
-                <geometry>
-                    <box size="${base_size_x*0.7} ${base_size_y*0.7} ${laser_slit_height}" />
-                </geometry>
-                <material name="LightGrey"/>
-            </visual>
-            <visual>
-                <origin xyz="0 0 ${wheel_slit_height + base_size_z_bottom + laser_slit_height + base_size_z_top/2}" rpy="0 0 0"/>
-                <geometry>
-                    <box size="${base_size_x} ${base_size_y} ${base_size_z_top}" />
-                </geometry>
-                <material name="DarkGrey"/>
-            </visual>
-
-            <!-- base bottom collisions -->
-            <collision>
-                <origin xyz="0 0 ${wheel_slit_height + base_size_z_bottom/2}" rpy="0 0 0.0"/>
-                <geometry>
-                    <box size="${base_size_x} ${base_size_y} ${base_size_z_bottom}" />
-                </geometry>
-            </collision>
-
-            <!-- base top collisions -->
-            <collision>
-                <origin xyz="0 0 ${wheel_slit_height + base_size_z_bottom + laser_slit_height + base_size_z_top/2}" rpy="0 0 0.0"/>
-                <geometry>
-                    <box size="${base_size_x} ${base_size_y} ${base_size_z_top}" />
-                </geometry>
-            </collision>
-
-            <!-- <xacro:cuboid_inertia mass="${base_mass}" length="${base_size_x}" -->
-                                 <!-- breadth="${base_size_y}" height="${base_size_z}"-->     
-        </link>
-
-        <joint name="${name}_joint" type="fixed">
-            <xacro:insert_block name="origin" />
-            <parent link="${parent}"/>
-            <child link="${name}_link" />
-        </joint>
-
-        <!-- WHEELS -->
-        <xacro:kelo_drive name="kelo_drive_front_left" parent="${name}_link" >
-            <origin xyz="${base_size_x*0.5 - wheel_radius}
-                         ${base_size_y*0.5 - wheel_radius}
-                         ${wheel_radius}"
-                    rpy="0 0 0"/>
-        </xacro:kelo_drive>
-        <xacro:kelo_drive name="kelo_drive_front_right" parent="${name}_link" >
-            <origin xyz="${base_size_x*0.5 - wheel_radius}
-                         ${-(base_size_y*0.5 - wheel_radius)}
-                         ${wheel_radius}"
-                    rpy="0 0 0"/>
-        </xacro:kelo_drive>
-        <xacro:kelo_drive name="kelo_drive_rear_right" parent="${name}_link" >
-            <origin xyz="${-(base_size_x*0.5 - wheel_radius)}
-                         ${base_size_y*0.5 - wheel_radius}
-                         ${wheel_radius}"
-                    rpy="0 0 0"/>
-        </xacro:kelo_drive>
-        <xacro:kelo_drive name="kelo_drive_rear_left" parent="${name}_link" >
-            <origin xyz="${-(base_size_x*0.5 - wheel_radius)}
-                         ${-(base_size_y*0.5 - wheel_radius)}
-                         ${wheel_radius}"
-                    rpy="0 0 0"/>
-        </xacro:kelo_drive>
-
-    </xacro:macro>
-</robot>
