+++ eddie_base_description/urdf/eddie_robot.urdf.xacro
@@ -1,0 +1,12 @@
+<?xml version="1.0"?>
+<robot xmlns:xacro="http://www.ros.org/wiki/xacro" name="eddie_base_with_wheels">
+
+    <!-- Include the Eddie Base -->
+    <xacro:include filename="$(find eddie_base_description)/urdf/eddie_base.macro.xacro"/>
+
+    <xacro:eddie_with_no_arm/>
+
+    
+
+
+</robot>
