--- eddie_base_description/urdf/materials.urdf.xacro
@@ -1,30 +1,0 @@
-<?xml version="1.0"?>
-<robot>
-  <material name="Orange">
-    <color rgba="1.0 0.4 0.0 1.0" />
-  </material>
-  <material name="Grey">
-    <color rgba="0.7 0.7 0.7 1.0" />
-  </material>
-  <material name="DarkGrey">
-    <color rgba="0.3 0.3 0.3 1.0" />
-  </material>
-  <material name="White">
-    <color rgba="1.0 1.0 1.0 1.0" />
-  </material>
-  <material name="Black">
-    <color rgba="0.0 0.0 0.0 1.0" />
-  </material>
-  <material name="Red">
-    <color rgba="1.0 0.0 0.0 1.0" />
-  </material>
-  <material name="Green">
-    <color rgba="0.0 1.0 0.0 1.0" />
-  </material>
-  <material name="Blue">
-    <color rgba="0.0 0.0 0.8 1.0" />
-  </material>
-  <material name="LightGrey">
-    <color rgba="0.82 0.82 0.82 1.0" />
-  </material>
-</robot>
