--- eddie_description/CMakeLists.txt
@@ -1,18 +1,0 @@
-cmake_minimum_required(VERSION 3.12)
-project(eddie_description)
-
-find_package(ament_cmake REQUIRED)
-
-install(DIRECTORY config
-        DESTINATION share/${PROJECT_NAME}
-)
-
-install(DIRECTORY robots
-        DESTINATION share/${PROJECT_NAME}
-)
-
-install(DIRECTORY launch
-        DESTINATION share/${PROJECT_NAME}
-)
-
-ament_package()
