--- eddie_description/config/initial_positions_arm_left.yaml
@@ -1,8 +1,0 @@
-# NOTE: Used for both 6 and 7 dof bots. The joint_7 value will be unused for 6 dof bots
-joint_1: -0.70 #-0.517734469311598
-joint_2: -2.05 #-0.08483200000000002
-joint_3: 2.14 #1.7743715307475147
-joint_4: -2.44 #-1.4395920000000002
-joint_5: -1.57 #0.0
-joint_6: 0.00 #-0.708248
-joint_7: 0.0
