--- eddie_description/config/initial_positions_arm_right.yaml
@@ -1,8 +1,0 @@
-# Initial position of the right arm
-joint_1: 0.70
-joint_2: -2.05
-joint_3: 0.90
-joint_4: 2.44
-joint_5: 1.57
-joint_6: 0.00
-joint_7: 0.0
