--- eddie_description/config/ros2_controllers.yaml
@@ -1,57 +1,0 @@
-controller_manager:
-  ros__parameters:
-    update_rate: 1000  # Hz
-
-    joint_state_broadcaster:
-      type: joint_state_broadcaster/JointStateBroadcaster
-
-    joint_trajectory_controller:
-      type: joint_trajectory_controller/JointTrajectoryController
-
-    twist_controller:
-      type: picknik_twist_controller/PicknikTwistController
-
-    robotiq_gripper_controller:
-      type: position_controllers/GripperActionController
-
-    fault_controller:
-      type: picknik_reset_fault_controller/PicknikResetFaultController
-
-joint_trajectory_controller:
-  ros__parameters:
-    joints:
-      - kinova_left_joint_2
-      - kinova_left_joint_1
-      - kinova_left_joint_3
-      - kinova_left_joint_4
-      - kinova_left_joint_5
-      - kinova_left_joint_6
-      - kinova_left_joint_7
-    command_interfaces:
-      - position
-    state_interfaces:
-      - position
-      - velocity
-    state_publish_rate: 100.0
-    action_monitor_rate: 20.0
-    allow_partial_joints_goal: false
-    constraints:
-      stopped_velocity_tolerance: 0.0
-      goal_time: 0.0
-
-twist_controller:
-  ros__parameters:
-    joint: tcp
-    interface_names:
-      - twist.linear.x
-      - twist.linear.y
-      - twist.linear.z
-      - twist.angular.x
-      - twist.angular.y
-      - twist.angular.z
-
-robotiq_gripper_controller:
-  ros__parameters:
-    default: true
-    joint: robotiq_85_left_knuckle_joint
-    allow_stalling: true
