--- eddie_description/robots/kortex_robot.xacro
@@ -1,109 +1,0 @@
-<?xml version="1.0"?>
-
-<robot xmlns:xacro="http://ros.org/wiki/xacro">
-
-  <xacro:macro name="kinova_arm"
-    params="
-    parent
-    *origin
-    prefix
-    arm
-    gripper
-    gripper_joint_name
-    dof
-    vision
-    robot_ip
-    username
-    password
-    port
-    port_realtime
-    session_inactivity_timeout_ms
-    connection_inactivity_timeout_ms
-    use_internal_bus_gripper_comm:=false
-    use_fake_hardware:=false
-    fake_sensor_commands:=false
-    sim_gz:=false
-    sim_isaac:=false
-    isaac_joint_commands:=/isaac_joint_commands
-    isaac_joint_states:=/isaac_joint_states
-    use_external_cable:=false
-    initial_positions:=${dict(joint_1=0.0,joint_2=0.0,joint_3=0.0,joint_4=0.0,joint_5=0.0,joint_6=0.0,joint_7=0.0)}
-    gripper_max_velocity:=100.0
-    gripper_max_force:=100.0
-    gripper_com_port:=/dev/ttyUSB0
-    moveit_active:=false">
-
-    <!-- Include and load arm macro files -->
-    <xacro:include
-      filename="$(find kortex_description)/arms/${arm}/${dof}dof/urdf/${arm}_macro.xacro" />
-
-    <!-- Load the arm -->
-    <xacro:load_arm
-      parent="${parent}"
-      dof="${dof}"
-      vision="${vision}"
-      robot_ip="${robot_ip}"
-      username="${username}"
-      password="${password}"
-      port="${port}"
-      port_realtime="${port_realtime}"
-      session_inactivity_timeout_ms="${session_inactivity_timeout_ms}"
-      connection_inactivity_timeout_ms="${connection_inactivity_timeout_ms}"
-      prefix="${prefix}"
-      use_internal_bus_gripper_comm="${use_internal_bus_gripper_comm}"
-      use_fake_hardware="${use_fake_hardware}"
-      fake_sensor_commands="${fake_sensor_commands}"
-      sim_isaac="${sim_isaac}"
-      isaac_joint_commands="${isaac_joint_commands}"
-      isaac_joint_states="${isaac_joint_states}"
-      gripper_joint_name="${gripper_joint_name}"
-      gripper_max_velocity="${gripper_max_velocity}"
-      gripper_max_force="${gripper_max_force}"
-      use_external_cable="${use_external_cable}"
-      initial_positions="${initial_positions}"
-      moveit_active="${moveit_active}">
-      <xacro:insert_block name="origin" />
-    </xacro:load_arm>
-
-    <!-- If no gripper, define tool frame here -->
-    <xacro:if value="${not gripper}">
-      <link name="${prefix}tool_frame" />
-      <joint name="${prefix}tool_frame_joint" type="fixed">
-        <origin xyz="0 0 0" rpy="0 0 0" />
-        <parent link="${prefix}${last_arm_link}" />
-        <child link="${prefix}tool_frame" />
-        <axis xyz="0 0 0" />
-      </joint>
-    </xacro:if>
-
-    <!-- Include and load the gripper if defined -->
-    <xacro:unless value="${not gripper}">
-      <xacro:include filename="$(find robotiq_description)/urdf/robotiq_2f_85_macro.urdf.xacro" />
-
-      <!-- Hardware talks directly to the gripper so we don't need ros2_control unless we are
-      simulating -->
-      <xacro:property name="include_ros2_control" value="false" />
-      <xacro:if
-        value="${sim_gz or use_fake_hardware or not use_internal_bus_gripper_comm}">
-        <xacro:property name="include_ros2_control" value="true" />
-      </xacro:if>
-
-      <xacro:robotiq_gripper
-        name="RobotiqGripperHardwareInterface"
-        prefix="${prefix}"
-        parent="${prefix}${last_arm_link}"
-        include_ros2_control="false"
-        com_port="${gripper_com_port}"
-        use_fake_hardware="${use_fake_hardware}"
-        mock_sensor_commands="${fake_sensor_commands}"
-        sim_isaac="${sim_isaac}"
-        isaac_joint_commands="${isaac_joint_commands}"
-        isaac_joint_states="${isaac_joint_states}">
-        <origin xyz="0 0 0" rpy="0 0 0" />
-      </xacro:robotiq_gripper>
-
-    </xacro:unless>
-
-  </xacro:macro>
-
-</robot>
