--- eddie_torso_description/CMakeLists.txt
@@ -1,10 +1,0 @@
-cmake_minimum_required(VERSION 3.12)
-project(eddie_torso_description)
-
-find_package(ament_cmake REQUIRED)
-
-install(DIRECTORY meshes urdf
-  DESTINATION share/${PROJECT_NAME}
-)
-
-ament_package()
