--- eddie_torso_description/package.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>eddie_torso_description</name>
-  <version>0.1.0</version>
-  <description>Description of the torso of Eddie in form of URDF files and meshes</description>
-
-  <maintainer email="robotics@inf.h-brs.de">MAS robotics</maintainer>
-
-  <license>GPLv3</license>
-
-  <author email="aleksandar.mitrevski@h-brs.de">Alex Mitrevski</author>
-  <author email="dharmingeo@gmail.com">Dharmin B.</author>
-  <author email="sushant.chavan@smail.inf.h-brs.de">Sushant Chavan</author>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
