--- eddie_torso_description/urdf/eddie_torso_gz.urdf.xacro
@@ -1,147 +1,0 @@
-<?xml version="1.0"?>
-<robot xmlns:xacro="http://www.ros.org/wiki/xacro">
-
-    <!-- torso parameters -->
-    <xacro:property name="torso_height" value="0.3" />
-    <xacro:property name="torso_width" value="0.4" />
-    <xacro:property name="torso_thickness" value="0.19" />
-    <xacro:property name="torso_back_offset" value="0.02" />
-
-    <xacro:property name="torso_shoulder_z_offset" value="0.03" />
-    <xacro:property name="torso_shoulder_height" value="0.2" />
-    <xacro:property name="torso_shoulder_width" value="0.2" />
-    <xacro:property name="torso_shoulder_thickness" value="0.02" />
-
-    <xacro:include filename="$(find eddie_base_description)/urdf/common.xacro" />
-    <xacro:include filename="$(find eddie_base_description)/urdf/eddie_base_gz.urdf.xacro" />
-
-    <xacro:macro name="eddie_torso" params="name parent *origin use_gpu_2d_laser:=true">
-        <!-- TORSO -->
-        <joint name="${name}_joint" type="fixed">
-            <origin
-                xyz="${-base_size_x/2 + torso_back_offset + torso_thickness/2} 0 ${wheel_slit_height + base_size_z}"
-                rpy="0 0 0" />
-            <parent link="${parent}" />
-            <child link="${name}_link" />
-        </joint>
-
-        <link name="${name}_link">
-            <visual>
-                <origin
-                    xyz="0 0 ${torso_height/2}"
-                    rpy="0 0 0" />
-                <geometry>
-                    <box size="${torso_thickness} ${torso_width} ${torso_height}" />
-                </geometry>
-                <material name="Grey"/>
-            </visual>
-
-            <!-- torso collision -->
-            <collision>
-                <origin
-                    xyz="0 0 ${torso_height/2}"
-                    rpy="0 0 0" />
-                <geometry>
-                    <box size="${torso_thickness} ${torso_width} ${torso_height}" />
-                </geometry>
-            </collision>
-
-            <xacro:cuboid_inertia mass="${base_mass/10000}" length="${base_size_x}"
-                breadth="${base_size_y}" height="${base_size_z}" />
-        </link>
-
-        <!-- TORSO SHOULDER LEFT -->
-        <joint name="${name}_shoulder_left_joint" type="fixed">
-            <origin
-                xyz="${torso_thickness/2 - torso_shoulder_thickness/2} 0 ${torso_height - torso_shoulder_z_offset}"
-                rpy="0. 0.1309 0.7853982" />
-            <parent link="${name}_link" />
-            <child link="${name}_shoulder_left_link" />
-        </joint>
-
-        <link name="${name}_shoulder_left_link">
-            <visual>
-                <origin
-                    xyz="0 ${torso_shoulder_width/2} ${torso_shoulder_height/2}"
-                    rpy="0 0 0" />
-                <geometry>
-                    <box
-                        size="${torso_shoulder_thickness} ${torso_shoulder_width} ${torso_shoulder_height}" />
-                </geometry>
-                <material name="DarkGrey"/>
-            </visual>
-
-            <!-- torso shoulder collision -->
-            <collision>
-                <origin
-                    xyz="0 ${torso_shoulder_width/2} ${torso_shoulder_height/2}"
-                    rpy="0 0 0" />
-                <geometry>
-                    <box
-                        size="${torso_shoulder_thickness} ${torso_shoulder_width} ${torso_shoulder_height}" />
-                </geometry>
-            </collision>
-
-            <xacro:cuboid_inertia mass="${base_mass}" length="${base_size_x}"
-                breadth="${base_size_y}" height="${base_size_z}" />
-        </link>
-
-        <joint name="${name}_shoulder_left_arm_joint" type="fixed">
-            <origin
-                xyz="0 ${torso_shoulder_width/2} ${torso_shoulder_height/2}"
-                rpy="3.14159 -1.5708 0" />
-            <parent link="${name}_shoulder_left_link" />
-            <child link="${name}_shoulder_left_arm_attach_link" />
-        </joint>
-
-        <link name="${name}_shoulder_left_arm_attach_link" />
-
-        <!-- TORSO SHOULDER RIGHT -->
-        <joint name="${name}_shoulder_right_joint" type="fixed">
-            <origin
-                xyz="${torso_thickness/2 - torso_shoulder_thickness/2} 0 ${torso_height - torso_shoulder_z_offset}"
-                rpy="0 0.1309 -0.7853982" />
-            <parent link="${name}_link" />
-            <child link="${name}_shoulder_right_link" />
-        </joint>
-
-        <link name="${name}_shoulder_right_link">
-            <visual>
-                <origin
-                    xyz="0 ${-torso_shoulder_width/2} ${torso_shoulder_height/2}"
-                    rpy="0 0 0" />
-                <geometry>
-                    <box
-                        size="${torso_shoulder_thickness} ${torso_shoulder_width} ${torso_shoulder_height}" />
-                </geometry>
-                <material name="DarkGrey"/>
-            </visual>
-
-            <!-- torso shoulder collision -->
-            <collision>
-                <origin
-                    xyz="0 ${-torso_shoulder_width/2} ${torso_shoulder_height/2}"
-                    rpy="0 0 0" />
-                <geometry>
-                    <box
-                        size="${torso_shoulder_thickness} ${torso_shoulder_width} ${torso_shoulder_height}" />
-                </geometry>
-            </collision>
-            
-            <!-- This is causing the base to topple over under combined x y velocity -->
-            <xacro:cuboid_inertia mass="${base_mass}" length="${base_size_x}"
-                breadth="${base_size_y}" height="${base_size_z}" /> 
-        </link>
-
-        <joint name="${name}_shoulder_right_arm_joint" type="fixed">
-            <origin
-                xyz="0 ${-torso_shoulder_width/2} ${torso_shoulder_height/2}"
-                rpy="3.14159 -1.5708 0" />
-            <parent link="${name}_shoulder_right_link" />
-            <child link="${name}_shoulder_right_arm_attach_link" />
-        </joint>
-
-         <link name="${name}_shoulder_right_arm_attach_link" />
-
-    </xacro:macro>
-</robot>
